Move DateOfBirth class into its own file
